refactor(prompt-area): fold paste-resolution back into engine
resolveTriggersInSegments + splitTextByTriggerPatterns were extracted
into their own 110-line file during Phase A, but the module had exactly
one caller (use-prompt-area-events.ts:handlePaste) and one export —
classic orphan-utility smell. Also, isValidTriggerPosition and
mergeAdjacentTextSegments (both engine exports) are the only
dependencies, so the code sits most naturally adjacent to them.

Moves both functions back into prompt-area-engine.ts next to the chip
lifecycle section. Deletes prompt-area-paste-resolution.ts. Updates the
import in use-prompt-area-events.ts and the engine test. Drops the
registry.json entry, brings registry.test.ts file count 19 → 18.

While here:
- Collapses redundant `displayText || query` fallback after
  `onSelect?.() ?? query` into a single `|| query`. Treats '' from
  onSelect as "no custom label" (an empty chip renders blank). Picks
  one contract instead of mixing ?? and || (kieran-ts K9).
- Drops the dead `i === 0` check in splitTextByTriggerPatterns's inner
  while loop — unreachable since we only enter after `i++` (kieran-ts
  K10).

Engine: 547 → 659 lines. 657/657 tests pass.

Surfaced by ce:review (reviewers: maintainability, code-simplicity).
