+++ registry.json
@@ -35,10 +35,6 @@
           "type": "registry:component"
         },
         {
-          "path": "registry/new-york/blocks/prompt-area/prompt-area-paste-resolution.ts",
-          "type": "registry:component"
-        },
-        {
           "path": "registry/new-york/blocks/prompt-area/cursor-helpers.ts",
           "type": "registry:component"
         },
